+++ examples/streaming/wordcount-java/src/main/java/org/apache/gearpump/streaming/examples/wordcountjava/dsl/refactor/WordCount.java
@@ -29,8 +29,8 @@ import org.apache.gearpump.cluster.client.ClientContext;
 import org.apache.gearpump.streaming.dsl.api.functions.ReduceFunction;
 import org.apache.gearpump.streaming.hadoop.HadoopCheckpointStoreFactory;
 import org.apache.gearpump.streaming.hadoop.lib.rotation.FileSizeRotation;
-import org.apache.gearpump.streaming.refactor.coder.StringUtf8Coder;
-import org.apache.gearpump.streaming.refactor.coder.VarLongCoder;
+import org.apache.gearpump.streaming.coder.StringUtf8Coder;
+import org.apache.gearpump.streaming.coder.VarLongCoder;
 import org.apache.gearpump.streaming.refactor.dsl.api.functions.MapWithStateFunction;
 import org.apache.gearpump.streaming.refactor.dsl.javaapi.JavaStream;
 import org.apache.gearpump.streaming.dsl.javaapi.functions.GroupByFunction;
