+++ streaming/src/main/java/org/apache/gearpump/streaming/coder/CoderException.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.gearpump.streaming.refactor.coder;
+package org.apache.gearpump.streaming.coder;
 
 public class CoderException extends RuntimeException {
     public CoderException(String message) {
