+++ streaming/src/main/java/org/apache/gearpump/streaming/coder/SetCoder.java
@@ -16,30 +16,34 @@
  * limitations under the License.
  */
 
-package org.apache.gearpump.streaming.refactor.coder;
+package org.apache.gearpump.streaming.coder;
 
-import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.List;
-import java.util.Observer;
+import java.util.Set;
 
-public abstract class ElementByteSizeObservableIterable<
-        V, InputT extends ElementByteSizeObservableIterator<V>>
-        implements Iterable<V> {
-    private List<Observer> observers = new ArrayList<>();
+public class SetCoder<T> extends IterableLikeCoder<T, Set<T>> {
 
-    protected abstract InputT createIterator();
-
-    public void addObserver(Observer observer) {
-        observers.add(observer);
+    public static <T> SetCoder<T> of(Coder<T> elementCoder) {
+        return new SetCoder<>(elementCoder);
     }
 
     @Override
-    public InputT iterator() {
-        InputT iterator = createIterator();
-        for (Observer observer : observers) {
-            iterator.addObserver(observer);
-        }
-        observers.clear();
-        return iterator;
+    public void verifyDeterministic() throws NonDeterministicException {
+        throw new NonDeterministicException(this,
+                "Ordering of elements in a set may be non-deterministic.");
     }
+
+    /////////////////////////////////////////////////////////////////////////////
+    // Internal operations below here.
+
+    @Override
+    protected final Set<T> decodeToIterable(List<T> decodedElements) {
+        return new HashSet<>(decodedElements);
+    }
+
+    protected SetCoder(Coder<T> elemCoder) {
+        super(elemCoder, "Set");
+    }
+
 }
