+++ streaming/src/main/scala/org/apache/gearpump/streaming/refactor/state/heap/HeapStateInternalsProxy.scala
@@ -18,7 +18,7 @@
 
 package org.apache.gearpump.streaming.refactor.state.heap
 
-import org.apache.gearpump.streaming.refactor.coder.Coder
+import org.apache.gearpump.streaming.coder.Coder
 import org.apache.gearpump.streaming.refactor.state.{StateNamespace, StateTag}
 import org.apache.gearpump.streaming.refactor.state.api.{State, StateInternals, StateInternalsFactory}
